Tue, Sep 5, 2023, 7:22 PM +02:00

+++ readability/readability.c
@@ -25,7 +25,8 @@ int main(void)
             sentences++;
         }
     }
-    float L = letters / words * 100;
-    float S = semtences / words * 100;
+    float L = (float) letters / (float) words * 100;
+    float S = (float)semtences / (float)words * 100;
     int index = round (0.0588 * L - 0.296 * S - 15.8);
+    
 }
